requirements.txt
boto3==1.13.16
botocore==1.16.16
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
idna==2.9
jmespath==0.10.0
lxml==4.5.1
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
